// Frontend/client/src/api/models/PipelineDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PipelineDTO
 */
export interface PipelineDTO {
    /**
     * Unique identifier of the pipeline
     * @type {number}
     * @memberof PipelineDTO
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof PipelineDTO
     */
    pipelineData: string;
    /**
     * User who uploaded the image
     * @type {string}
     * @memberof PipelineDTO
     */
    uploadedBy: string;
    /**
     * Timestamp for when the pipeline was created
     * @type {Date}
     * @memberof PipelineDTO
     */
    createdAt: Date;
}

/**
 * Check if a given object implements the PipelineDTO interface.
 */
export function instanceOfPipelineDTO(value: object): value is PipelineDTO {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('pipelineData' in value) || value['pipelineData'] === undefined) return false;
    if (!('uploadedBy' in value) || value['uploadedBy'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    return true;
}

export function PipelineDTOFromJSON(json: any): PipelineDTO {
    return PipelineDTOFromJSONTyped(json, false);
}

export function PipelineDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): PipelineDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'pipelineData': json['pipelineData'],
        'uploadedBy': json['uploaded_by'],
        'createdAt': (new Date(json['created_at'])),
    };
}

export function PipelineDTOToJSON(json: any): PipelineDTO {
    return PipelineDTOToJSONTyped(json, false);
}

export function PipelineDTOToJSONTyped(value?: PipelineDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'pipelineData': value['pipelineData'],
        'uploaded_by': value['uploadedBy'],
        'created_at': ((value['createdAt']).toISOString()),
    };
}